+++ thicket-front/src/payment/pages/Payment.jsx
@@ -1,7 +1,9 @@
-import React, {useEffect, useState} from 'react';
+import React, {useEffect, useRef, useState} from 'react';
 import 'bootstrap/dist/css/bootstrap.css';
 import ReadyKakaopay from "../../member/pages/ReadyKakaopay";
-import Bank from "../component/Bank";
+import useDetectClose from "../component/UseDetectClose";
+import {Dropdown} from "../component/Dropdown";
+import "../../assets/css/style.css"
 
 
 function Payment({ selectedChair, selectedTime, selectedDate, selectedQuantity }) {
@@ -9,6 +11,31 @@ function Payment({ selectedChair, selectedTime, selectedDate, selectedQuantity }
     const [show, setShow] = useState([]);       // 공연정보
     const [times, setTimes] = useState([]);     // 공연정보-시간리스트
     const [chairs, setChairs] = useState([]);   // 단일시간-좌석리스트
+    const dropDownRef = useRef();
+    const [identify, setIdentify] = useState('선택해주세요');
+    const [generatedNumber, setGeneratedNumber] = useState(null);
+    const list = ['농협은행', '국민은행', '하나은행'];
+    const [isOpen, setIsOpen] = useDetectClose(dropDownRef, false);
+
+    const generateNumber = (bank) => {
+        switch (bank) {
+            case '농협은행':
+                return '12345';
+            case '국민은행':
+                return '45678';
+            case '하나은행':
+                return '11111';
+            default:
+                return '';
+        }
+    };
+
+    const handleBankSelection = (selectedBank) => {
+        const generatedNum = generateNumber(selectedBank);
+        setGeneratedNumber(generatedNum);
+        setIdentify(selectedBank);
+        setIsOpen(false);
+    };
 
     useEffect(() => {
         // 페이지가 처음 렌더링될 때 초기 상태
@@ -16,26 +43,27 @@ function Payment({ selectedChair, selectedTime, selectedDate, selectedQuantity }
     }, []);
 
     useEffect(() => { // 공연정보 (stage.stage uuid)
-        fetch('/shows/stagedetail/51f864ca-1352-4b9b-80fe-359ad340c136')
+        fetch('/shows/stagedetail/c2e8bbaa-5c12-4cb7-930a-33552c85d81f')
             .then(response => response.json())
             .then(data => {
                 setShow(data);
             });
     }, []);
     useEffect(() => { // 공연정보 - 시간리스트 (stage.stage uuid)
-        fetch('/tickets/all/51f864ca-1352-4b9b-80fe-359ad340c136')
+        fetch('/tickets/all/c2e8bbaa-5c12-4cb7-930a-33552c85d81f')
             .then(response => response.json())
             .then(data => {
                 setTimes(data);
             });
     }, []);
     useEffect(() => { // 단일시간 - 좌석리스트 (stage.stage_start uuid)
-        fetch('/chairs/all/892c4364-a8f0-4e58-85ef-41d2f6434331')
+        fetch('/chairs/all/2a1b5854-00a9-4b9e-8017-33fa3b0429f9')
             .then(response => response.json())
             .then(data => {
                 setChairs(data);
             });
     }, []);
+
 
     return (
         <>
@@ -51,6 +79,7 @@ function Payment({ selectedChair, selectedTime, selectedDate, selectedQuantity }
                     alignItems: 'center',
                     position: 'relative',
                     borderRadius: '12px',
+                    fontFamily: 'mont'
                 }}
             >
                 <div style={{ width: '19%' }}>
@@ -83,19 +112,23 @@ function Payment({ selectedChair, selectedTime, selectedDate, selectedQuantity }
                     alignItems: 'center',
                     height: '80vh',
                     backgroundColor: 'white',
+                    fontFamily: 'mont'
                 }}
             >
                 <div style={{ padding: '5px', maxHeight: '80vh', width: '31%', textAlign: 'left'}}>
                     <h1 style={{borderBottom: '1.5px solid black', fontSize: '15px'}}>티켓 수령방법</h1>
-                    <div className="form-check" style={{ backgroundColor: 'lightgray', borderRadius: '5px', height: '55px', padding: '5px'}}>
+                    <div className="form-check" style={{ backgroundColor: 'lightgray', borderRadius: '5px',
+                                                            height: '55px', padding: '5px'}}>
                         <div style={{ display: 'flex', flexDirection: 'row', marginBottom: '10px'}}>
-                        <input className="form-check-input" type="radio" name="flexRadioDefault1" id="flexRadioDefault2" checked />
+                        <input className="form-check-input" type="radio" name="flexRadioDefault1"
+                                                            id="flexRadioDefault2" checked />
                         <label className="form-check-label" htmlFor="flexRadioDefault2">
                             현장수령
                         </label>
                         </div>
                         <div style={{ display: 'flex', flexDirection: 'row'}}>
-                        <input className="form-check-input" type="radio" name="flexRadioDefault1" id="flexRadioDefault1" disabled />
+                        <input className="form-check-input" type="radio" name="flexRadioDefault1"
+                                                                id="flexRadioDefault1" disabled />
                         <label className="form-check-label" htmlFor="flexRadioDefault1">
                             모바일 티켓
                         </label>
@@ -104,7 +137,8 @@ function Payment({ selectedChair, selectedTime, selectedDate, selectedQuantity }
                     <br />
                     <br />
                     <h1 style={{borderBottom: '1.5px solid black', fontSize: '15px'}}>결제 방식 선택</h1>
-                    <div className="form-check" style={{ backgroundColor: 'lightgray', borderRadius: '5px', height: '105px', padding: '5px' }} >
+                    <div className="form-check" style={{ backgroundColor: 'lightgray', borderRadius: '5px',
+                                                            height: '105px', padding: '5px' }} >
 
                         <div style={{ display: 'flex', flexDirection: 'row', marginBottom: '5px' }}>
                         <input className="form-check-input"
@@ -155,14 +189,58 @@ function Payment({ selectedChair, selectedTime, selectedDate, selectedQuantity }
                 >
                 </div>
                 <div style={{ padding: '5px', width: '33%', textAlign: 'center', justifyContent: 'center'}}>
-                    {/*<h1 style={{fontSize: '20px'}}>예매가 정상적으로 완료되었습니다.</h1>*/}
-                    {/*<p> 다음날 자정까지 결제하지 않으면 어쩌구 저쩌구 됩니다!!!</p>*/}
                     <h1 style={{ fontSize: '15px'}}>
                         {paymentMethod === 'kakao'
                         ? <ReadyKakaopay />
-
                             :
-                                <Bank />
+                            <div style={{ padding: '10px', maxHeight: '80vh', width: '90%', textAlign: 'left' }}>
+                                <p style={{ padding: '6px', borderBottom: '2px solid darkgray' }}>
+                                    결제 수단 입력
+                                </p>
+                                <p style={{ width: '93%', backgroundColor: 'lightgray', padding: '6px',
+                                    borderRadius: '5px', marginTop: '-13px', marginLeft: '1px'}}> > 무통장 입금 </p>
+                                <table style={{ borderBottom: '2px solid darkgray', marginTop: '-15px' }}>
+                                    <tbody>
+                                    <tr style={{ height: '110px' }}>
+                                        <th style={{ textAlign: 'center', backgroundColor: 'lightgray',
+                                            width: '31%', borderRadius: '5px' }}>
+                                            입금액<br /><br />
+                                            <p style={{ fontSize: '13', padding: '3px' }}>입금 하실 은행</p>
+                                        </th>
+                                        <td style={{ textAlign: 'right', padding: '8px' }}>
+                                            {Number(selectedChair.price * selectedQuantity).toLocaleString()}원
+                                            <br /><br />
+
+                                            <div ref={dropDownRef}>
+                                                <button
+                                                    onClick={() => setIsOpen(!isOpen)}
+                                                    style={{ padding: '5px', cursor: 'pointer',
+                                                        border: 'none' ,borderRadius: '5px'}}
+                                                >
+                                                    {identify}
+                                                </button>
+                                                {isOpen &&
+                                                    <div style={{cursor: 'pointer'}}>
+                                                        {list.map((value) => (
+                                                            <Dropdown
+                                                                key={value}
+                                                                value={value}
+                                                                setIsOpen={setIsOpen}
+                                                                setIdentify={handleBankSelection}
+                                                                isOpen={isOpen} />
+                                                        ))}
+                                                    </div>
+                                                }
+                                            </div>
+                                            {generatedNumber &&
+                                                <div style={{ marginTop: '5px', fontSize: '15px' }}>
+                                                    계좌번호: {generatedNumber}
+                                                </div>}
+                                        </td>
+                                    </tr>
+                                    </tbody>
+                                </table>
+                            </div>
                             }
                     </h1>
                 </div>
@@ -178,12 +256,12 @@ function Payment({ selectedChair, selectedTime, selectedDate, selectedQuantity }
                     <table>
                         <tbody>
                         <tr style={{height: '150px'}}>
-                            <th style={{ textAlign: 'center', backgroundColor: 'lightgray', width: '25%', borderRadius: '5px'}}>
+                            <th style={{ textAlign: 'center', backgroundColor: 'lightgray', width: '27%', borderRadius: '5px'}}>
                                 공연 <br /><br /><br />
                                 장소 <br /><br />
                                 좌석
                             </th>
-                            <td style={{ textAlign: 'center', width: '80%'}}>
+                            <td style={{ textAlign: 'center', width: '80%', padding: '15px'}}>
                                 {show.name}<br /><br />
                                 {show.place} <br /><br />
                                 {selectedChair && selectedChair.chairType}석 {selectedQuantity}매
