--- thicket-member/src/test/resources/secretKey.properties
@@ -1,1 +1,0 @@
-jwtSecret: cGxheWRhdGEtZmluYWwtcHJvamVjdC10aGlja2V0LWNlcnRpZnkta2V5IGJhc2U=
